[#19] refactor : 학교 인증 컴포넌트 분할
